// src/types.ts
export interface NavLink {
  label: string;
  href: string;
}

export interface StatItem {
  number: string;
  label: string;
}

export interface EngStatItem {
  label: string;
  value: string;
}

export interface BootColorOption {
  id: string;
  name: string;
  colorClass: string;
  imageSrc: string;
}